#5 - Moved password hasher and password criteria validator
