+++ server/dependencies/password_hasher.go
@@ -1,18 +1,18 @@
 package dependencies
 
 import (
-	"blendermux/server/controllers"
+	"blendermux/server/helpers"
 	"sync"
 )
 
 var passwordHasherOnce sync.Once
-var passwordHasher controllers.PasswordHasher
+var passwordHasher helpers.PasswordHasher
 
 // ResolvePasswordHasher resolves the PasswordHasher dependency.
 // Only the first call to this function will create a new PasswordHasher, after which it will be retrieved from the cache.
-func ResolvePasswordHasher() controllers.PasswordHasher {
+func ResolvePasswordHasher() helpers.PasswordHasher {
 	passwordHasherOnce.Do(func() {
-		passwordHasher = controllers.BCryptPasswordHasher{}
+		passwordHasher = helpers.BCryptPasswordHasher{}
 	})
 	return passwordHasher
 }
