+++ server/helpers/password_criteria_validator.go
@@ -1,4 +1,4 @@
-package models
+package helpers
 
 // PasswordCriteriaValidator is an interface for validating a password against criteria.
 type PasswordCriteriaValidator interface {
