+++ server/helpers/password_hasher.go
@@ -1,4 +1,4 @@
-package controllers
+package helpers
 
 import (
 	"blendermux/common"
